+++ App/PersonalHygiene/Features/Settings/Views/DiagnosticsViewSections.swift
@@ -138,6 +138,7 @@ extension DiagnosticsView {
                 authTimelineSection
                 networkActivitySection
                 pendingDetailsSection
+                pendingByGroupSection
                 Section {
                     Button {
                         Task {